+++ frontend/techpick-extension/src/apis/apiClient.ts
@@ -1,7 +1,7 @@
+import { ERROR_MESSAGE_JSON } from '@/constants/errorMessageJson';
+import { notifyError } from '@/libs/@toast/notifyError';
+import { returnErrorFromHTTPError } from '@/utils/returnErrorFromHTTPError';
 import ky, { HTTPError } from 'ky';
-import { notifyError } from '@/libs/@toast';
-import { returnErrorFromHTTPError } from '@/utils';
-import { ERROR_MESSAGE_JSON } from '@/constants';
 
 export const apiClient = ky.create({
   credentials: 'include',
