+++ frontend/techpick-extension/src/apis/getRootFolderChildFolders.ts
@@ -1,13 +1,13 @@
+import { API_URLS } from '@/constants/apiUrls';
+import type { GetRootFolderChildFoldersResponseType } from '@/types/GetRootFolderChildFoldersResponseType';
+import { returnErrorFromHTTPError } from '@/utils/returnErrorFromHTTPError';
 import { HTTPError } from 'ky';
-import { returnErrorFromHTTPError } from '@/utils';
-import { API_URLS } from '@/constants';
-import type { GetRootFolderChildFoldersResponseType } from '@/types';
 import { apiClient } from './apiClient';
 
 export const getRootFolderChildFolders = async () => {
   try {
     const response = await apiClient.get<GetRootFolderChildFoldersResponseType>(
-      API_URLS.getFoldersUrl()
+      API_URLS.getFoldersUrl(),
     );
     const data = await response.json();
     return data;
