+++ frontend/techpick-extension/src/apis/getTagList.ts
@@ -1,13 +1,13 @@
-import { apiClient } from './apiClient';
-import type { GetTagListResponseType } from '@/types';
-import { API_URLS } from '@/constants';
+import { API_URLS } from '@/constants/apiUrls';
+import type { GetTagListResponseType } from '@/types/GetTagListResponseType';
+import { returnErrorFromHTTPError } from '@/utils/returnErrorFromHTTPError';
 import { HTTPError } from 'ky';
-import { returnErrorFromHTTPError } from '@/utils';
+import { apiClient } from './apiClient';
 
 export const getTagList = async () => {
   try {
     const response = await apiClient.get<GetTagListResponseType>(
-      API_URLS.getTagsUrl()
+      API_URLS.getTagsUrl(),
     );
     const data = await response.json();
