--- frontend/techpick-extension/src/apis/index.ts
@@ -1,9 +1,0 @@
-export { createTag } from './createTag';
-export { deleteTag } from './deleteTag';
-export { checkPickByUrl } from './checkPickByUrl';
-export { getTagList } from './getTagList';
-export { updatePickFromExtension } from './updatePickFromExtension';
-export { updateTag } from './updateTag';
-export { getBasicFolderList } from './getBasicFolders';
-export { getRootFolderChildFolders } from './getRootFolderChildFolders';
-export { createPickToUnclassifiedFolder } from './createPickToUnclassifiedFolder';
