+++ frontend/techpick-extension/src/apis/updatePickFromExtension.ts
@@ -1,21 +1,19 @@
-import { API_URLS } from '@/constants';
-import { apiClient } from './apiClient';
-import type {
-  UpdatePickFromExtensionRequestType,
-  UpdatePickFromExtensionResponseType,
-} from '@/types';
+import { API_URLS } from '@/constants/apiUrls';
+import type { UpdatePickFromExtensionRequestType } from '@/types/UpdatePickFromExtensionRequestType';
+import type { UpdatePickFromExtensionResponseType } from '@/types/UpdatePickFromExtensionResponseType';
+import { returnErrorFromHTTPError } from '@/utils/returnErrorFromHTTPError';
 import { HTTPError } from 'ky';
-import { returnErrorFromHTTPError } from '@/utils';
+import { apiClient } from './apiClient';
 
 export const updatePickFromExtension = async (
-  pickInfo: UpdatePickFromExtensionRequestType
+  pickInfo: UpdatePickFromExtensionRequestType,
 ) => {
   try {
     const response = await apiClient.patch<UpdatePickFromExtensionResponseType>(
       API_URLS.getExtensionPickUrl(),
       {
         json: pickInfo,
-      }
+      },
     );
     const data = await response.json();
