+++ frontend/techpick-extension/src/apis/updateTag.ts
@@ -1,8 +1,8 @@
-import { UpdateTagRequestType } from '@/types';
-import { apiClient } from './apiClient';
-import { API_URLS } from '@/constants';
+import { API_URLS } from '@/constants/apiUrls';
+import type { UpdateTagRequestType } from '@/types/UpdateTagRequestType';
+import { returnErrorFromHTTPError } from '@/utils/returnErrorFromHTTPError';
 import { HTTPError } from 'ky';
-import { returnErrorFromHTTPError } from '@/utils';
+import { apiClient } from './apiClient';
 
 export const updateTag = async (updateTagInfo: UpdateTagRequestType) => {
   try {
